smart-breakout-strategy/requirements.txt
python-binance
pandas
ta
python-telegram-bot


